clientdataをinternalに移動
