+++ CMakeLists.txt
@@ -158,12 +158,17 @@ set(SVR_SRC
 )
 set(CLI_SRC
     ${CPP_DIR}/client/client.cc
-    ${CPP_DIR}/client/client_data.cc
     ${CPP_DIR}/client/client_ws.cc
+    ${CPP_DIR}/client/data_store1.cc
+    ${CPP_DIR}/client/data_store2.cc
     ${CPP_DIR}/client/field.cc
     ${CPP_DIR}/client/func.cc
+    ${CPP_DIR}/client/log.cc
     ${CPP_DIR}/client/logger.cc
     ${CPP_DIR}/client/member.cc
+    ${CPP_DIR}/client/text.cc
+    ${CPP_DIR}/client/value.cc
+    ${CPP_DIR}/client/view.cc
 )
 set(EXAMPLE_SRC
     ${CPP_DIR}/example/main.cc
