+++ src/include/webcface/func.h
@@ -1,17 +1,20 @@
 #pragma once
 #include <vector>
-#include "client_data.h"
 #include "common/func.h"
 #include "common/val.h"
 #include "func_result.h"
 #include "common/def.h"
 
 namespace WebCFace {
+namespace Internal {
+struct ClientData;
+}
 
 namespace FuncWrapper {
 
 //! Client::sync() まで待機し、実行完了までsync()をブロックするFuncWrapper
-WEBCFACE_DLL FuncWrapperType runCondOnSync(const std::weak_ptr<ClientData> &data);
+WEBCFACE_DLL FuncWrapperType
+runCondOnSync(const std::weak_ptr<Internal::ClientData> &data);
 //! ScopeGuardをロックするFuncWrapper
 template <typename ScopeGuard>
 inline FuncWrapperType runCondScopeGuard() {
@@ -33,7 +36,7 @@ class Func : protected Field {
     friend AnonymousFunc;
 
     Func() = default;
-    Func(const Field &base) : Field(base) {}
+    WEBCFACE_DLL Func(const Field &base);
     Func(const Field &base, const std::string &field)
         : Func(Field{base, field}) {}
 
@@ -41,12 +44,12 @@ class Func : protected Field {
     using Field::name;
 
   private:
-    auto &set(const std::shared_ptr<FuncInfo> &v) {
-        setCheck();
-        dataLock()->func_store.setSend(*this, v);
-        return *this;
+    WEBCFACE_DLL Func &setRaw(const std::shared_ptr<FuncInfo> &v);
+    Func &setRaw(const FuncInfo &v) {
+        return setRaw(std::make_shared<FuncInfo>(v));
     }
-    auto &set(const FuncInfo &v) { return set(std::make_shared<FuncInfo>(v)); }
+    WEBCFACE_DLL std::optional<std::shared_ptr<FuncInfo>> getRaw() const;
+    WEBCFACE_DLL FuncWrapperType getDefaultFuncWrapper() const;
 
   public:
     //! 関数からFuncInfoを構築しセットする
@@ -55,28 +58,22 @@ class Func : protected Field {
      * 引数や戻り値の情報などは更新されない
      */
     template <typename T>
-    auto &set(const T &func) {
-        return this->set(
-            FuncInfo{std::function{func}, dataLock()->default_func_wrapper});
+    Func &set(const T &func) {
+        return this->setRaw(
+            FuncInfo{std::function{func}, getDefaultFuncWrapper()});
     }
     //! 関数からFuncInfoを構築しセットする
     template <typename T>
-    auto &operator=(const T &func) {
+    Func &operator=(const T &func) {
         return this->set(func);
     }
 
     //! 関数を関数リストで非表示にする
     //! (他clientのentryに表示されなくする)
-    auto &hidden(bool hidden) {
-        setCheck();
-        dataLock()->func_store.setHidden(*this, hidden);
-        return *this;
-    }
+    WEBCFACE_DLL Func &hidden(bool hidden);
+
     //! 関数の設定を削除
-    auto &free() {
-        dataLock()->func_store.unsetRecv(*this);
-        return *this;
-    }
+    WEBCFACE_DLL Func &free();
 
     //! 関数を実行する (同期)
     /*! selfの関数の場合、このスレッドで直接実行する
@@ -112,10 +109,22 @@ class Func : protected Field {
     runAsync(const std::vector<ValAdaptor> &args_vec) const;
 
     //! 戻り値の型を返す
-    WEBCFACE_DLL ValType returnType() const;
+    WEBCFACE_DLL ValType returnType() const {
+        auto func_info = getRaw();
+        if (func_info) {
+            return (*func_info)->return_type;
+        }
+        return ValType::none_;
+    }
     //! 引数の情報を返す
     //! 変更するにはsetArgsを使う(このvectorの中身を書き換えても反映されない)
-    WEBCFACE_DLL std::vector<Arg> args() const;
+    WEBCFACE_DLL std::vector<Arg> args() const {
+        auto func_info = getRaw();
+        if (func_info) {
+            return (*func_info)->args;
+        }
+        return std::vector<Arg>{};
+    }
     const Arg args(std::size_t i) const { return args().at(i); }
 
     //! 引数の情報を更新する
@@ -157,11 +166,8 @@ class Func : protected Field {
 };
 
 
-class WEBCFACE_DLL AnonymousFunc : public Func {
-    static std::string fieldNameTmp() {
-        static int id = 0;
-        return ".tmp" + std::to_string(id++);
-    }
+class AnonymousFunc : public Func {
+    WEBCFACE_DLL static std::string fieldNameTmp();
 
     std::function<void(AnonymousFunc &)> func_setter = nullptr;
     bool base_init = false;
@@ -182,15 +188,6 @@ class WEBCFACE_DLL AnonymousFunc : public Func {
         };
     }
 
-    void lockTo(Func &target) {
-        if (!base_init) {
-            this->data_w = target.data_w;
-            this->member_ = target.member_;
-            this->field_ = fieldNameTmp();
-            func_setter(*this);
-        }
-        target.set(dataLock()->func_store.getRecv(*this).value());
-        this->free();
-    }
+    WEBCFACE_DLL void lockTo(Func &target);
 };
 } // namespace WebCFace
