+++ src/include/webcface/member.h
@@ -5,15 +5,17 @@
 #include "field.h"
 #include "event_target.h"
 #include "common/def.h"
+#include "value.h"
+#include "text.h"
+#include "log.h"
+#include "func.h"
+#include "view.h"
 
 namespace WebCFace {
 
-class Value;
-class Text;
-class Func;
-class Log;
-class View;
+namespace Internal {
 struct ClientData;
+}
 
 //! Memberを指すクラス
 /*!
@@ -23,7 +25,7 @@ struct ClientData;
 class WEBCFACE_DLL Member : protected Field {
   public:
     Member() = default;
-    Member(const std::weak_ptr<ClientData> &data_w, const std::string &member)
+    Member(const std::weak_ptr<Internal::ClientData> &data_w, const std::string &member)
         : Field(data_w, member) {}
     Member(const Field &base) : Field(base) {}
