+++ src/include/webcface/text.h
@@ -5,31 +5,31 @@
 #include <memory>
 #include "common/dict.h"
 #include "field.h"
-#include "client_data.h"
 #include "event_target.h"
 
 namespace WebCFace {
+namespace Internal {
+struct ClientData;
+}
+class Member;
 
 //! 文字列の送受信データを表すクラス
 /*! コンストラクタではなく Member::text() を使って取得してください
  */
 class Text : protected Field, public EventTarget<Text> {
+  public:
+    using Dict = Common::Dict<std::shared_ptr<std::string>>;
+
+  private:
+    WEBCFACE_DLL Text &set(const std::shared_ptr<std::string> &v);
+    WEBCFACE_DLL std::optional<std::shared_ptr<std::string>> getRaw() const;
+    WEBCFACE_DLL std::optional<Dict> getRawRecurse() const;
 
     void onAppend() const override { tryGet(); }
 
-    // set(Dict)の内部でつかう
-    Text &set(const std::shared_ptr<std::string> &v) {
-        setCheck();
-        dataLock()->text_store.setSend(*this, v);
-        this->triggerEvent(*this);
-        return *this;
-    }
-
   public:
     Text() = default;
-    Text(const Field &base)
-        : Field(base), EventTarget<Text>(&this->dataLock()->text_change_event,
-                                         *this) {}
+    WEBCFACE_DLL Text(const Field &base);
     Text(const Field &base, const std::string &field)
         : Text(Field{base, field}) {}
 
@@ -44,7 +44,6 @@ class Text : protected Field, public EventTarget<Text> {
         return Text{*this, this->field_ + "." + field};
     }
 
-    using Dict = Common::Dict<std::shared_ptr<std::string>>;
     //! Dictの値を再帰的にセットする
     Text &set(const Dict &v) {
         if (v.hasValue()) {
@@ -73,7 +72,7 @@ class Text : protected Field, public EventTarget<Text> {
 
     //! 文字列を返す
     std::optional<std::string> tryGet() const {
-        auto v = dataLock()->text_store.getRecv(*this);
+        auto v = getRaw();
         if (v) {
             return **v;
         } else {
@@ -81,9 +80,7 @@ class Text : protected Field, public EventTarget<Text> {
         }
     }
     //! 文字列を再帰的に取得しDictで返す
-    std::optional<Dict> tryGetRecurse() const {
-        return dataLock()->text_store.getRecvRecurse(*this);
-    }
+    std::optional<Dict> tryGetRecurse() const { return getRawRecurse(); }
     //! 値を返す
     std::string get() const { return tryGet().value_or(""); }
     //! 値を再帰的に取得しDictで返す
@@ -91,11 +88,7 @@ class Text : protected Field, public EventTarget<Text> {
     operator std::string() const { return get(); }
     operator Dict() const { return getRecurse(); }
     //! syncの時刻を返す
-    std::chrono::system_clock::time_point time() const {
-        return dataLock()
-            ->sync_time_store.getRecv(this->member_)
-            .value_or(std::chrono::system_clock::time_point());
-    }
+    WEBCFACE_DLL std::chrono::system_clock::time_point time() const;
 
     // //! このtext非表示にする
     // //! (他clientのentryに表示されなくする)
@@ -106,17 +99,13 @@ class Text : protected Field, public EventTarget<Text> {
     // }
 
     //! 値やリクエスト状態をクリア
-    Text &free() {
-        dataLock()->text_store.unsetRecv(*this);
-        return *this;
-    }
+    WEBCFACE_DLL Text &free();
 
     bool operator==(const std::string &rhs) const { return this->get() == rhs; }
     bool operator!=(const std::string &rhs) const { return this->get() != rhs; }
 };
 
-inline std::ostream &operator<<(std::ostream &os,
-                                            const Text &data) {
+inline std::ostream &operator<<(std::ostream &os, const Text &data) {
     return os << data.get();
 }
 } // namespace WebCFace
