+++ src/include/webcface/value.h
@@ -4,32 +4,34 @@
 #include <chrono>
 #include <memory>
 #include "common/dict.h"
+#include "common/vector.h"
 #include "field.h"
-#include "client_data.h"
 #include "event_target.h"
 
 namespace WebCFace {
+namespace Internal {
+struct ClientData;
+}
+class Member;
 
 //! 実数値またはその配列の送受信データを表すクラス
-/*! コンストラクタではなく Member::value(), Member::values(), Member::onValueEntry() を使って取得してください
+/*! コンストラクタではなく Member::value(), Member::values(),
+ * Member::onValueEntry() を使って取得してください
  */
 class Value : protected Field, public EventTarget<Value> {
+public:
+    using Dict = Common::Dict<std::shared_ptr<Common::VectorOpt<double>>>;
+    private:
 
-    void onAppend() const override { tryGet(); }
+    WEBCFACE_DLL Value &set(const std::shared_ptr<VectorOpt<double>> &v);
+    WEBCFACE_DLL std::optional<std::shared_ptr<VectorOpt<double>>> getRaw() const;
+    WEBCFACE_DLL std::optional<Dict> getRawRecurse() const;
 
-    // set(Dict)の内部でつかう
-    Value &set(const std::shared_ptr<VectorOpt<double>> &v) {
-        setCheck();
-        dataLock()->value_store.setSend(*this, v);
-        this->triggerEvent(*this);
-        return *this;
-    }
+    void onAppend() const override { getRaw(); }
 
   public:
     Value() = default;
-    Value(const Field &base)
-        : Field(base), EventTarget<Value>(&this->dataLock()->value_change_event,
-                                          *this) {}
+    WEBCFACE_DLL Value(const Field &base);
     Value(const Field &base, const std::string &field)
         : Value(Field{base, field}) {}
 
@@ -44,7 +46,6 @@ class Value : protected Field, public EventTarget<Value> {
         return Value{*this, this->field_ + "." + field};
     }
 
-    using Dict = Common::Dict<std::shared_ptr<Common::VectorOpt<double>>>;
     //! Dictの値を再帰的にセットする
     Value &set(const Dict &v) {
         if (v.hasValue()) {
@@ -75,7 +76,7 @@ class Value : protected Field, public EventTarget<Value> {
 
     //! 値を返す
     std::optional<double> tryGet() const {
-        auto v = dataLock()->value_store.getRecv(*this);
+        auto v = getRaw();
         if (v) {
             return **v;
         } else {
@@ -84,7 +85,7 @@ class Value : protected Field, public EventTarget<Value> {
     }
     //! 値をvectorで返す
     std::optional<std::vector<double>> tryGetVec() const {
-        auto v = dataLock()->value_store.getRecv(*this);
+        auto v = getRaw();
         if (v) {
             return **v;
         } else {
@@ -93,7 +94,7 @@ class Value : protected Field, public EventTarget<Value> {
     }
     //! 値を再帰的に取得しDictで返す
     std::optional<Dict> tryGetRecurse() const {
-        return dataLock()->value_store.getRecvRecurse(*this);
+        return getRawRecurse();
     }
     //! 値を返す
     double get() const { return tryGet().value_or(0); }
@@ -107,17 +108,10 @@ class Value : protected Field, public EventTarget<Value> {
     operator std::vector<double>() const { return getVec(); }
     operator Dict() const { return getRecurse(); }
     //! syncの時刻を返す
-    std::chrono::system_clock::time_point time() const {
-        return dataLock()
-            ->sync_time_store.getRecv(this->member_)
-            .value_or(std::chrono::system_clock::time_point());
-    }
+    WEBCFACE_DLL std::chrono::system_clock::time_point time() const;
 
     //! 値やリクエスト状態をクリア
-    Value &free() {
-        dataLock()->value_store.unsetRecv(*this);
-        return *this;
-    }
+    WEBCFACE_DLL Value &free();
 
     Value &operator+=(double rhs) {
         this->set(this->get() + rhs);
@@ -187,8 +181,7 @@ class Value : protected Field, public EventTarget<Value> {
 };
 
 
-inline std::ostream &operator<<(std::ostream &os,
-                                            const Value &data) {
+inline std::ostream &operator<<(std::ostream &os, const Value &data) {
     return os << data.get();
 }
